% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdnoticeshipment}
\alias{ecs_rdnoticeshipment}
\title{发货通知处理}
\usage{
ecs_rdnoticeshipment(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
发货通知处理
}
\examples{
ecs_rdnoticeshipment()
}
