% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdpurchasesbilling}
\alias{ecs_rdpurchasesbilling}
\title{应付单数据处理}
\usage{
ecs_rdpurchasesbilling(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
应付单数据处理
}
\examples{
ecs_rdpurchasesbilling()
}
