% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdreceiptnotice}
\alias{ecs_rdreceiptnotice}
\title{收料通知处理}
\usage{
ecs_rdreceiptnotice(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
收料通知处理
}
\examples{
ecs_rdreceiptnotice()
}
