% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdreturnnotice}
\alias{ecs_rdreturnnotice}
\title{其他出库处理}
\usage{
ecs_rdreturnnotice(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
其他出库处理
}
\examples{
ecs_rdreturnnotice()
}
