% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdreturnpurchase}
\alias{ecs_rdreturnpurchase}
\title{采购退料处理}
\usage{
ecs_rdreturnpurchase(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
采购退料处理
}
\examples{
ecs_rdreturnpurchase()
}
