% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdreturnrequest}
\alias{ecs_rdreturnrequest}
\title{退货申请处理}
\usage{
ecs_rdreturnrequest(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
退货申请处理
}
\examples{
ecs_rdreturnrequest()
}
