% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdreturnsales}
\alias{ecs_rdreturnsales}
\title{销售退货处理}
\usage{
ecs_rdreturnsales(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
销售退货处理
}
\examples{
ecs_rdreturnsales()
}
