% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_rdsalesbilling}
\alias{ecs_rdsalesbilling}
\title{同步应收单处理}
\usage{
ecs_rdsalesbilling(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
同步应收单处理
}
\examples{
ecs_rdsalesbilling()
}
