% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECSpack.R
\name{ecs_saleOder}
\alias{ecs_saleOder}
\title{销售订单处理}
\usage{
ecs_saleOder(startDate, endDate)
}
\arguments{
\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
销售订单处理
}
\examples{
ecs_saleOder()
}
